Rename deprecated folder for consistency
