+++ libs/@guardian/source/src/foundations/__deprecated__/typography/data.ts
