+++ libs/@guardian/source/src/foundations/__deprecated__/typography/typographyDeprecated.mdx
